Revert "move log output into LogTrait"
This reverts commit b784703da881e5b4c6553c577ee2b7ca6d3e17ef.
